moved country data into own file
