+++ src/data/countries.js
@@ -1,0 +1,21 @@
+var countries = [
+  {
+    name: 'New Zealand',
+    eligible: false,
+    age: 60,
+    hasCitizenship: true
+  },
+  {
+    name: 'Israel',
+    eligible: false,
+    age: 62,
+    hasCitizenship: true
+  },
+  {
+    name: 'Uruguay',
+    eligible: true,
+    hasCitizenship: true
+  }
+];
+
+export default countries;
